layout/forms/test/test_bug411236.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=411236
-->
<head>
  <title>Test for Bug 411236</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=411236">Mozilla Bug 411236</a>
<p id="display"></p>
<div id="content">
  <input type="file" onfocus="window.oTarget = event.originalTarget;"
         onclick="window.fileInputGotClick = true; return false;"
         id="fileinput">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 411236 **/

window.oTarget = null;
window.fileInputGotClick = false;

// accessibility.tabfocus must be set to value 7 before running test also
// on a mac.
function setOrRestoreTabFocus(newValue) {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  const prefSvcContractID = "@mozilla.org/preferences-service;1";
  const prefSvcIID = Components.interfaces.nsIPrefService;
  var prefs = Components.classes[prefSvcContractID].getService(prefSvcIID)
                                                   .getBranch("accessibility.");
  if (!newValue) {
    prefs.clearUserPref("tabfocus");
  } else {
    prefs.setIntPref("tabfocus", newValue);
  }
}

function tab() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var utils = window.QueryInterface(Components.interfaces.nsIInterfaceRequestor).
                     getInterface(Components.interfaces.nsIDOMWindowUtils);
  // Send tab key events.
  var key = Components.interfaces.nsIDOMKeyEvent.DOM_VK_TAB;
  utils.sendKeyEvent("keydown", key, 0, 0);
  utils.sendKeyEvent("keypress", key, 0, 0);
  utils.sendKeyEvent("keyup", key, 0, 0);
}

function test() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  // Try to find the 'Browse...' using tabbing.
  var i = 0;
  while (!window.oTarget && i < 100) {
    ++i;
    tab();
  }

  if (i >= 100) {
    ok(false, "Couldn't find an input element!");
    setOrRestoreTabFocus(0);
    SimpleTest.finish();
    return;
  }

  ok(window.oTarget instanceof HTMLInputElement, "Should have focused an input element!")
  ok(window.oTarget.type == "button", "Should have focused 'Browse...' button!");
  var e = document.createEvent("mouseevents");
  e.initMouseEvent("click", true, true, window, 0, 1, 1, 1, 1,
                    false, false, false, false, 0, null);
  window.oTarget.dispatchEvent(e);
  ok(window.fileInputGotClick,
     "File input should have got a click event, but not open the file dialog.");
  setOrRestoreTabFocus(0);
  SimpleTest.finish();
}

function do_test() {
  setOrRestoreTabFocus(7);
  window.focus();
  document.getElementById('fileinput').focus();
  setTimeout(test, 100);
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(do_test);

</script>
</pre>
</body>
</html>
